-spojene inzeraty s usermi

+++ SEMKA_mvc/App/Views/Inzerat/detail.view.php
@@ -1,4 +1,9 @@
-<?php /** @var Array $data */ ?>
+<?php /** @var Array $data */
+/** @var \App\Core\AAuthenticator $auth */
+
+use App\Models\User;
+
+?>
 <h1 >Aktuálny inzerát</h1>
 <hr/>
 
@@ -16,11 +21,13 @@
             <img src="<?= $data->getObrazok() ?>" alt="" class="img-fluid">
         </div>
     </div>
-
+    <p >Pridal : <?= $data->getOwner()->getLogin() ?></p>
     <hr/>
+
 
    <div class="icena">
         <p>Cena : <?= $data->getCena() ?> €</p>
+
    </div>
 
 </div>
@@ -30,9 +37,13 @@
 
 <div class="container-fluid">
     <div class="row">
+        <?php if ($auth->isLogged()  && ($auth->getLoggedUser()->getId() == $data->getIdOwner())){ ?>
+            <a href="?c=Inzerat&a=Edit&id=<?= $data->getId() ?>" class="btn tlacidlo edit col-6">Editovať inzerát</a>
+            <a href="?c=Inzerat&a=Delete&id=<?= $data->getId() ?>" class="btn tlacidlo danger col-6">Vymazať inzerát</a>
 
-        <a href="?c=Inzerat&a=Edit&id=<?= $data->getId() ?>" class="btn tlacidlo edit col-6">Editovať inzerát</a>
-        <a href="?c=Inzerat&a=Delete&id=<?= $data->getId() ?>" class="btn tlacidlo danger col-6">Vymazať inzerát</a>
+        <?php } ?>
+
+
